### 2025-05-27 3dce099
Create CONTRIBUTING.md
