+++ CONTRIBUTING.md
@@ -1,0 +1,172 @@
+# Contributing to JaxLayerLumos
+
+First off, thanks for taking the time to contribute to JaxLayerLumos! ❤️
+
+All types of contributions are encouraged and valued. This document provides guidelines for contributing to JaxLayerLumos. Please make sure to read the relevant section before making your contribution. It will make it a lot easier for us maintainers and smooth out the experience for all involved. The community looks forward to your contributions. 🎉
+
+> And if you like the project, but just don't have time to contribute, that's fine. There are other easy ways to support the project and show your appreciation, which we would also be very happy about:
+> - Star the project on GitHub!
+> - Tweet about JaxLayerLumos
+> - Refer this project in your project's readme or publications
+> - Mention the project at local meetups and tell your friends/colleagues
+
+## Table of Contents
+
+- [Code of Conduct](#code-of-conduct)
+- [I Have a Question](#i-have-a-question)
+- [I Want To Contribute](#i-want-to-contribute)
+  - [Reporting Bugs](#reporting-bugs)
+  - [Suggesting Enhancements](#suggesting-enhancements)
+  - [Your First Code Contribution](#your-first-code-contribution)
+  - [Improving The Documentation](#improving-the-documentation)
+- [Styleguides](#styleguides)
+  - [Commit Messages](#commit-messages)
+  - [Python Code Style](#python-code-style)
+- [Development Setup](#development-setup)
+
+## Code of Conduct
+
+This project and everyone participating in it is governed by the [JaxLayerLumos Code of Conduct](CODE_OF_CONDUCT.md). By participating, you are expected to uphold this code. Please report unacceptable behavior to [YOUR_EMAIL_OR_PROJECT_EMAIL@example.com].
+
+*(Note: You will need to create a `CODE_OF_CONDUCT.md` file. You can adapt a template like the [Contributor Covenant](https://www.contributor-covenant.org/version/2/1/code_of_conduct.html))*
+
+## I Have a Question
+
+Before you ask a question, please ensure you have:
+- Read the available [Documentation for JaxLayerLumos](LINK_TO_YOUR_DOCUMENTATION_SITE_OR_README_SECTION).
+- Searched for existing [GitHub Issues](LINK_TO_JAXLAYERLUMOS_ISSUES) that might address your question.
+- Searched the internet for answers.
+
+If you still need to ask a question, we recommend the following:
+- Open an [Issue on GitHub](LINK_TO_JAXLAYERLUMOS_ISSUES/new).
+- Provide as much context as you can about what you're trying to achieve or what problem you're encountering.
+- Provide project and platform versions (e.g., Python version, JAX version, OS) if relevant.
+
+We will try to address your question as soon as possible.
+
+## I Want To Contribute
+
+> ### Legal Notice
+> When contributing to this project, you must agree that you have authored 100% of the content, that you have the necessary rights to the content, and that the content you contribute may be provided under the project's [MIT License](LINK_TO_YOUR_LICENSE_FILE).
+
+### Reporting Bugs
+
+#### Before Submitting a Bug Report
+
+A good bug report shouldn't leave others needing to chase you up for more information. Please investigate carefully, collect information, and describe the issue in detail.
+
+- **Ensure you are using the latest version** of JaxLayerLumos.
+- **Verify the bug:** Determine if it's a bug in JaxLayerLumos or an issue with your setup or usage (e.g., incompatible environment components/versions). Please consult the [documentation](LINK_TO_YOUR_DOCUMENTATION_SITE_OR_README_SECTION).
+- **Check existing issues:** Search the [bug tracker](LINK_TO_JAXLAYERLUMOS_ISSUES?q=is%3Aissue+label%3Abug) to see if the bug has already been reported.
[…]
