### 2025-05-27 d40a68e
Update paper.bib
