+++ paper.bib
@@ -4,6 +4,7 @@
   url = {http://github.com/jax-ml/jax},
   version = {0.3.13},
   year = {2018},
+  doi={10.5281/zenodo.10104100}
 }
 
 @article{Wang:15apr,
@@ -86,7 +87,8 @@
   title = {{tmm}: A transfer-matrix method (TMM) optics calculator},
   year = {2012},
   version = {0.2.0},
-  url = {https://github.com/sbyrnes321/tmm}
+  url = {https://github.com/sbyrnes321/tmm},
+  doi = {https://doi.org/10.48550/arXiv.1603.02720}
 }
 
 @article{luce2022tmm,
@@ -216,5 +218,6 @@
   title={Radar absorbing materials: from theory to design and characterization},
   author={Vinoy, Kalarickaparambil Joseph and Jha, Rakesh Mohan},
   year={1996},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/978-1-4613-0473-9}
 }
