+++ paper.bib
@@ -4,7 +4,6 @@
   url = {http://github.com/jax-ml/jax},
   version = {0.3.13},
   year = {2018},
-  doi={10.5281/zenodo.10104100}
 }
 
 @article{Wang:15apr,
@@ -88,7 +87,7 @@
   year = {2012},
   version = {0.2.0},
   url = {https://github.com/sbyrnes321/tmm},
-  doi = {https://doi.org/10.48550/arXiv.1603.02720}
+  doi = {10.48550/arXiv.1603.02720}
 }
 
 @article{luce2022tmm,
