+++ paper.bib
@@ -3,7 +3,7 @@
   title = {{JAX}: composable transformations of {P}ython+{N}um{P}y programs},
   url = {http://github.com/jax-ml/jax},
   version = {0.3.13},
-  year = {2018},
+  year = {2018}
 }
 
 @article{Wang:15apr,
