+++ paper.bib
@@ -82,8 +82,8 @@
 }
 
 @software{tmmSbyrnes,
-  author = {Steven Byrnes},
-  title = {{tmm}: A transfer-matrix method (TMM) optics calculator},
+  author = {Byrnes, Steven},
+  title = {{tmm}: Multilayer optical calculations},
   year = {2012},
   version = {0.2.0},
   url = {https://github.com/sbyrnes321/tmm},
