### 2025-06-23 #8280 @tomelse
Hi, this looks like a really interesting package, but I don't have time to review it unfortunately as I'm currently in the process of reviewing another JOSS paper.

Sorry about that!
