### 2025-07-11 #8572 @editorialbot
[REVIEW]: JaxLayerLumos: A JAX-based Differentiable Optical and Radio Frequency Simulator for Multilayer Structures
**Submitting author:** @Mil152 (<a href="http://orcid.org/0000-0001-6217-9382">Mingxuan  Li</a>)
**Repository:** https://github.com/JaxLayerLumos/JaxLayerLumos
**Branch with paper.md** (empty if default branch): JOSS
**Version:** v0.3.3
