### 2025-07-11 JaxLayerLumos/JaxLayerLumos#21 @arendMoerman
No instructions on running unittests
This issue is part of my JOSS review: https://github.com/openjournals/joss-reviews/issues/8572

The repository contains a folder named `tests` that contains a set of Python scripts, presumably containing (unit)tests.
To the best of my knowledge, the documentation does not mention how these tests can be run in an automated way.

Please add, at the very least, a sentence detailing how they can be run automatically, e.g. using `nose2` (requires the `nose2` package) or `pytest .` or something like that. I checked and both options work for me.

And please also mention that a user has to follow the dev install (`pip install .[dev]`) in order to run the unittests (the dependencies on `pytest` and `colour-science`). The current documentation states that:
```
...where dev is defined for installing the packages required for development...
```
This sentence is quite broad in scope (what are "packages required for development"?) and it might not be obvious that a dev install is required even for a normal user who just wants to, for example, run the tests to check the installation when installing from source.
