### 2025-07-11 JaxLayerLumos/JaxLayerLumos#22 @arendMoerman
Documentation missing API reference
This issue if part of my JOSS review: https://github.com/openjournals/joss-reviews/issues/8572

The documentation contains a lot of examples, which is nice, but I cannot find a section containing an API reference. The presence of core documentation, such as an API reference, is part of the review checklist and hence should be present. Please add one that, at the very least, documents the public interface.

As an example, please have a look at the [Scipy public API reference](https://docs.scipy.org/doc/scipy/reference/#api-definition).
