### 2025-07-11 JaxLayerLumos/JaxLayerLumos#23 @arendMoerman
Documentation missing community guidelines
This issue if part of my JOSS review: https://github.com/openjournals/joss-reviews/issues/8572

The documentation lacks a community guideline, which is part of the review checklist:

>**Community guidelines**: Are there clear guidelines for third parties wishing to 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support

Please add these guidelines to the documentation.
