### 2025-07-17 aeab02c
add docstrings
