+++ pyproject.toml
@@ -46,7 +46,6 @@ dev = [
     "build",
     "twine",
     "sphinx",
-    "sphinx_rtd_theme",
     "black",
     "colour-science",
 ]
