+++ docs/conf.py
@@ -6,6 +6,11 @@
 # -- Project information -----------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#project-information
 
+import os
+import sys
+
+sys.path.insert(0, os.path.abspath('../'))
+
 project = 'JaxLayerLumos'
 copyright = '2025, Mingxuan Li, Jungtaek Kim, and Paul W. Leu'
 author = 'Mingxuan Li, Jungtaek Kim, and Paul W. Leu'
@@ -16,13 +21,12 @@ release = 'v0.3.3'
 
 extensions = [
     'sphinx.ext.autodoc',
+    'sphinx.ext.napoleon',
     'sphinx_rtd_theme',
 ]
 
 templates_path = ['_templates']
 exclude_patterns = []
-
-
 
 # -- Options for HTML output -------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#options-for-html-output
