+++ docs/index.rst
@@ -1,17 +1,8 @@
-.. JaxLayerLumos documentation master file, created by
-   sphinx-quickstart on Fri Jul 25 16:11:57 2025.
-   You can adapt this file completely to your liking, but it should at least
-   contain the root `toctree` directive.
-
-JaxLayerLumos documentation
-===========================
-
-Add your content using ``reStructuredText`` syntax. See the
-`reStructuredText <https://www.sphinx-doc.org/en/master/usage/restructuredtext/index.html>`_
-documentation for details.
-
+Documentation of JaxLayerLumos
+==============================
 
 .. toctree::
-   :maxdepth: 2
-   :caption: Contents:
+    :maxdepth: 2
+    :caption: Python API:
 
+    python_api/jaxlayerlumos
