+++ docs/python_api/jaxlayerlumos.rst
@@ -1,0 +1,8 @@
+jaxlayerlumos
+=============
+
+.. automodule:: jaxlayerlumos
+    :members:
+    :private-members:
+    :undoc-members:
+    :show-inheritance:
