+++ docs/python_api/jaxlayerlumos.jaxlayerlumos.rst
@@ -1,0 +1,8 @@
+jaxlayerlumos.jaxlayerlumos
+==========================
+
+.. automodule:: jaxlayerlumos.jaxlayerlumos
+    :members:
+    :private-members:
+    :undoc-members:
+    :show-inheritance:
