+++ docs/python_api/jaxlayerlumos.rst
@@ -6,3 +6,6 @@ jaxlayerlumos
     :private-members:
     :undoc-members:
     :show-inheritance:
+
+.. toctree::
+    jaxlayerlumos.jaxlayerlumos
