### 2025-07-27 784fe49
Update CONTRIBUTION.md and .readthedocs.yml
