+++ .readthedocs.yml
@@ -32,4 +32,6 @@ sphinx:
 # See https://docs.readthedocs.io/en/stable/guides/reproducible-builds.html
 python:
   install:
+    - method: pip
+      path: .[dev]
     - requirements: docs/requirements.txt
