+++ docs/python_api/jaxlayerlumos.colors.rst
@@ -1,0 +1,15 @@
+jaxlayerlumos.colors
+====================
+
+.. automodule:: jaxlayerlumos.colors
+    :members:
+    :private-members:
+    :undoc-members:
+    :show-inheritance:
+
+.. toctree::
+    jaxlayerlumos.colors.color_matching_functions
+    jaxlayerlumos.colors.composite
+    jaxlayerlumos.colors.illuminants
+    jaxlayerlumos.colors.transform
+    jaxlayerlumos.colors.utils
