+++ docs/python_api/jaxlayerlumos.jaxlayerlumos.rst
@@ -1,5 +1,5 @@
 jaxlayerlumos.jaxlayerlumos
-==========================
+===========================
 
 .. automodule:: jaxlayerlumos.jaxlayerlumos
     :members:
