+++ docs/python_api/jaxlayerlumos.colors.composite.rst
@@ -1,0 +1,8 @@
+jaxlayerlumos.colors.composite
+==============================
+
+.. automodule:: jaxlayerlumos.colors.composite
+    :members:
+    :private-members:
+    :undoc-members:
+    :show-inheritance:
