+++ docs/index.rst
@@ -1,8 +1,24 @@
 Documentation for JaxLayerLumos
 ===============================
 
+.. automodule:: python_api/jaxlayerlumos
+    :members:
+    :private-members:
+    :undoc-members:
+    :show-inheritance:
+
 .. toctree::
     :maxdepth: 2
     :caption: Python API:
 
-    python_api/jaxlayerlumos
+    python_api/jaxlayerlumos.jaxlayerlumos
+    python_api/jaxlayerlumos.utils_layers
+    python_api/jaxlayerlumos.utils_light_sources
+    python_api/jaxlayerlumos.utils_materials
+    python_api/jaxlayerlumos.utils_position
+    python_api/jaxlayerlumos.utils_radar_materials
+    python_api/jaxlayerlumos.utils_radio_frequency
+    python_api/jaxlayerlumos.utils_spectra
+    python_api/jaxlayerlumos.utils_units
+    python_api/jaxlayerlumos.wrappers
+    python_api/jaxlayerlumos.colors
