+++ docs/python_api/jaxlayerlumos.rst
@@ -1,21 +1,0 @@
-jaxlayerlumos
-=============
-
-.. automodule:: jaxlayerlumos
-    :members:
-    :private-members:
-    :undoc-members:
-    :show-inheritance:
-
-.. toctree::
-    jaxlayerlumos.jaxlayerlumos
-    jaxlayerlumos.utils_layers
-    jaxlayerlumos.utils_light_sources
-    jaxlayerlumos.utils_materials
-    jaxlayerlumos.utils_position
-    jaxlayerlumos.utils_radar_materials
-    jaxlayerlumos.utils_radio_frequency
-    jaxlayerlumos.utils_spectra
-    jaxlayerlumos.utils_units
-    jaxlayerlumos.wrappers
-    jaxlayerlumos.colors
