### 2025-07-27 JaxLayerLumos/JaxLayerLumos#21 @jungtaekkim
Hi @arendMoerman,

Thank you for your valuable review.

We have updated our `README.md` by considering your comment.

Please take a look at it:

https://github.com/JaxLayerLumos/JaxLayerLumos/commit/d109388c66996d37fbc5e64959cc27e4d917e854
