### 2025-07-27 JaxLayerLumos/JaxLayerLumos#23 @jungtaekkim
Hi @arendMoerman,

Thank you for your comment.

You can find a community guideline in the following link:

https://github.com/JaxLayerLumos/JaxLayerLumos/blob/JOSS/CONTRIBUTING.md
