### 2025-08-09 JaxLayerLumos/JaxLayerLumos#22 @arendMoerman
Thanks for adding an API reference. It looks good.

Howver, to the best of my knowledge, it cannot be found on the [main documentation pages](https://jaxlayerlumos.github.io/). Could you add a link to said documentation pages that clearly points towards the [API page you posted](https://jaxlayerlumos.readthedocs.io/en/joss)? Or if I am mistaken, could you show me where the link is?
