### 2025-08-09 JaxLayerLumos/JaxLayerLumos#23 @arendMoerman
Thanks for adding the CONTRIBUTING.md. I still miss it in the official documentation pages, which might make it a bit harder to find for users of your software. Could you please add a section in the landing page of your docs that point towards the contributing page, or that just states whatever you wrote in the CONTRIBUTING.md?
