### 2025-08-09 JaxLayerLumos/JaxLayerLumos#21 @arendMoerman
Thanks for adding this. Could you also please add it to the [documentation pages](https://jaxlayerlumos.github.io/)? To the best of my knowledge, I cannot find these instructions there.
