### 2025-08-09 JaxLayerLumos/JaxLayerLumos#24 @arendMoerman
Please spell out TPU in JOSS paper
This issue if part of my JOSS review: https://github.com/openjournals/joss-reviews/issues/8572

In the software paper, the acronym 'TPU' (tensor processing unit) is introduced at line 18. Could you please spell out this acronym before first usage? I understand things like 'GPU' are so commonplace nowadays (a google search immediately lands at 'graphics processing unit'), immediate usage of this acronym seems OK. But, for the general reader, the acronym 'TPU' might still be uncommon.
