### 2025-08-09 JaxLayerLumos/JaxLayerLumos#22 @jungtaekkim
@arendMoerman It is because we didn't merge the JOSS branch to the main branch.  Currently, the main branch does not have the documentation.  If a review process is done, we can make it accessible by merging the JOSS branch to the main branch.
