### 2025-08-10 JaxLayerLumos/JaxLayerLumos#25 @brandondube
Accuracy of performance claims
This issue if part of my JOSS review: https://github.com/openjournals/joss-reviews/issues/8572

I installed JaxLayerLumos and tmm_fast to compare performance.  I did these benchmarks on a desktop computer with an AMD Ryzen 7950x CPU, 32 GB of RAM, and an nvidia 4080 GPU, though I did not do any GPU-based benchmarking.  I edited your benchmark routine to remove ansys optics and only run JLL and tmm_fast.  The results are summarized in this table

| n layers | tmm_fast | jaxlayerlumos  | slowdown |
|----------|----------|----------------|----------|
| 2        | 0.0145   | 0.083          | 5.7x     |
| 4        | 0.0179   | 0.1009         | 5.6x     |
| 16       | 0.0345   | 0.2928         | 8.6x     |

Your software is several (nearly ten times!) times slower than the other software you include for comparison in benchmarking and label as "fast."

The benchmark script also emitted some messages:

> T_TE not matched: tmm_fast <-> jaxlayerlumos
> T_TM not matched: tmm_fast <-> jaxlayerlumos

If this gap were say 1 to 3x, I could accept the claim.  But at this sort of slowdown, I do not think this software can be called 'fast.'

I also note that the paper lists support for GPUs and TPUs, presumably on the basis that it is written in Jax and therefore supports these platforms.  However, these kinds of calculations can be sensitive to rounding type errors.  Since TPUs do not even support 32-bit numbers, I am skeptical a meaningful answer will come out of running this software on a TPU.  i.e., even if it technically runs, the result is not usable.

With a GPU, whether the code runs fast or slow depends on the number of operations done in parallel.  For example a problem size of 10,000 items is still likely to run faster on a CPU than GPU.

Could you please run the benchmark script (`python compare_methods.py`) edited lines 199-205

```
    methods = onp.array([
        # 'ansys',
        # 'tmm',
        'tmm_fast',
        # 'jaxlayerlumos_old',
        'jaxlayerlumos',
    ])
```

And reproduce my above table?  If you are able to run additional comparisons on GPU and TPU listed in a separate column, I am happy to look at that
