### 2025-08-10 JaxLayerLumos/JaxLayerLumos#23 @jungtaekkim
Thank you again.

You can find the updates regarding your request in these links:

https://jaxlayerlumos.github.io
https://github.com/JaxLayerLumos/JaxLayerLumos/blob/JOSS/README.md
