### 2025-08-10 ba120cd
Update paper.md
