### 2025-08-10 aea84b6
Update paper.md
