### 2025-08-18 d9d5ff0
Update paper.md
