### 2025-08-18 dd0db7b
Update paper.md
