### 2025-08-18 41520a3
Update paper.md
