### 2025-08-18 f2b8fb5
Update paper.md
