### 2025-08-19 JaxLayerLumos/JaxLayerLumos#25 @brandondube
Could you provide a timing result for each of ansys optics and tmm (not tmm fast) to compare?
