### 2025-08-25 JaxLayerLumos/JaxLayerLumos#25 @jungtaekkim
Hi @brandondube,

These are the timing results of the two implementations:

4 layers
ansys 4.5475 +- 0.5064 sec.
tmm 1.0224 +- 0.0202 sec.

8 layers
ansys 4.5380 +- 0.4639 sec.
tmm 1.6401 +- 0.0277 sec.

12 layers
ansys 4.5522 +- 0.4650 sec.
tmm 2.2409 +- 0.0400 sec.

16 layers
ansys 4.5164 +- 0.0477 sec.
tmm 2.8600 +- 0.0486 sec.

It is tested under the same setup described in [this file](https://github.com/JaxLayerLumos/JaxLayerLumos/blob/JOSS/benchmarking/compare_methods.py).
