### 2025-08-31 JaxLayerLumos/JaxLayerLumos#26 @brandondube
Lack of functional documentation
This issue if part of my JOSS review: https://github.com/openjournals/joss-reviews/issues/8572

The docs website is https://jaxlayerlumos.readthedocs.io/en/joss/

The API docs are available, but there is no "how to," "examples," or "tutorial" or other documentation showing a user or would-be user how to chain the functions together to perform any given calculation

Could you please add one of the above quoted items to the documentation?
