### 2025-09-04 7e277cd
Update run_on_cpus_or_tpus
