### 2025-09-05 JaxLayerLumos/JaxLayerLumos#25 @brandondube
Ok.  It seems the numeric accuracy issue is just as I expected.  Observe for example

```
CPU 0.57334185
TPU 0.58103985

diff = 0.007698 = 1.3%
```

For some, this may be an unacceptable error.  So, I would encourage noting that accuracy of calculations are reduced when using a TPU.
