### 2025-09-07 11f7adb
Update paper.md
