+++ paper/paper.md
@@ -114,6 +114,6 @@ Jupyter notebook examples are available in the [examples directory](https://gith
 
 # Acknowledgements
 
-We acknowledge the developers of JAX and other open-source libraries. This work was supported by the Center for Materials Data Science for Reliability and Degradation ([MDS-Rely](https://mds-rely.org/)), an Industry--University Cooperative Research Center of the National Science Foundation, through awards EEC-2052662, EEC-2052776, and EEC-2435402. We also thank the open-source community for contributions and feedback.
+We thank the open-source community for contributions and feedback. This work was supported by the Center for Materials Data Science for Reliability and Degradation ([MDS-Rely](https://mds-rely.org/)), an Industry--University Cooperative Research Center of the National Science Foundation (NSF), through NSF awards EEC-2052662, EEC-2052776, and EEC-2435402. Additionally, the University of Pittsburgh Center for Research Computing and Data (RRID:SCR_022735) provided essential resources for this work, supported by National Institutes of Health (NIH) award S10OD028483 and NSF award OAC-2117681.
 
 # References
