+++ docs/about/about_jaxlayerlumos.rst
@@ -1,0 +1,84 @@
+About JaxLayerLumos
+###################
+
+Overview
+========
+
+JaxLayerLumos is open-source transfer-matrix method (TMM) software designed for scientists, engineers, and researchers in optics and photonics. It provides a powerful yet intuitive interface for calculating the reflection and transmission (RT) of light through multi-layer optical structures. By inputting the refractive index, thickness of each layer, and the frequency vector, users can analyze how light interacts with layered materials, including the option to adjust for incidence angles.
+
+Our mission is to offer a lightweight, flexible, and fast alternative to commercial software, enabling users to perform complex optical simulations with ease. JaxLayerLumos is built with performance and usability in mind, facilitating the exploration of optical phenomena in research and development settings.
+
+Features
+========
+
+- Gradient Calculation: Calculates the gradients over any variables involved in RT, powered by JAX.
+- Flexibility: Accommodates a wide range of materials and structures by allowing users to specify complex refractive indices, layer thicknesses, and frequency vectors.
+- Angle of Incidence Support: Expands simulation capabilities to include angled light incidence, providing more detailed analysis for advanced optical designs.
+- Open Source and Community-Driven: Encourages contributions and feedback from the community, ensuring continuous improvement and innovation.
+- Comprehensive Material Database: Includes a growing database of materials with their optical properties, streamlining the simulation setup process.
+
+Installation
+============
+
+JaxLayerLumos can be easily installed by the following command using the [PyPI repository](https://pypi.org/project/jaxlayerlumos/).
+
+<code>pip install jaxlayerlumos</code><br><br>
+
+Alternatively, JaxLayerLumos can be installed from source.<br><br>
+
+<code>pip install .</code><br><br>
+
+In addition, we support three installation modes, <code>dev</code>, <code>benchmarking</code>, and <code>examples</code>, where <code>dev</code> is defined for installing the packages required for development and software testing, <code>benchmarking</code> is for installing the packages required for benchmarking against differnt TMM software programs, and <code>examples</code> is needed for running the examples included in the <code>examples</code> directory. One of these modes can be used by commanding <code>pip install .[dev]</code>, <code>pip install .[benchmarking]</code>, or <code>pip install .[examples]</code>.
+
+Examples
+========
+
+A collection of examples in the <code>examples</code> directory exhibits various use cases and capabilities of our software. We provide the following examples:
+
+1. [Reflection Spectra over Wavelengths Varying Incidence Angles](https://github.com/JaxLayerLumos/JaxLayerLumos/blob/main/examples/angle-variation.ipynb)
+2. [Color Conversion](https://github.com/JaxLayerLumos/JaxLayerLumos/blob/main/examples/color-conversion.ipynb)
+3. [Color Exploration with Thin-Film Structures](https://github.com/JaxLayerLumos/JaxLayerLumos/blob/main/examples/color-exploration.ipynb)
+4. [Gradient Computation](https://github.com/JaxLayerLumos/JaxLayerLumos/blob/main/examples/gradient-computation.ipynb)
+5. [Visualization of Light Sources](https://github.com/JaxLayerLumos/JaxLayerLumos/blob/main/examples/light-source-visualization.ipynb)
+
+6. [Plotting of Optical Constants](https://github.com/JaxLayerLumos/JaxLayerLumos/blob/main/examples/n-k-extrapolation.ipynb)
[…]
