+++ docs/about/about_jaxlayerlumos.rst
@@ -20,35 +20,37 @@ Features
 Installation
 ============
 
-JaxLayerLumos can be easily installed by the following command using the [PyPI repository](https://pypi.org/project/jaxlayerlumos/).
+JaxLayerLumos can be easily installed by the following command using the `PyPI repository <https://pypi.org/project/jaxlayerlumos/>`_.
 
-<code>pip install jaxlayerlumos</code><br><br>
+.. code-block:: bash
+    pip install jaxlayerlumos
 
-Alternatively, JaxLayerLumos can be installed from source.<br><br>
+Alternatively, JaxLayerLumos can be installed from source.
 
-<code>pip install .</code><br><br>
+.. code-block:: bash
+    pip install .
 
-In addition, we support three installation modes, <code>dev</code>, <code>benchmarking</code>, and <code>examples</code>, where <code>dev</code> is defined for installing the packages required for development and software testing, <code>benchmarking</code> is for installing the packages required for benchmarking against differnt TMM software programs, and <code>examples</code> is needed for running the examples included in the <code>examples</code> directory. One of these modes can be used by commanding <code>pip install .[dev]</code>, <code>pip install .[benchmarking]</code>, or <code>pip install .[examples]</code>.
+In addition, we support three installation modes, ``dev``, <code>benchmarking</code>, and <code>examples</code>, where <code>dev</code> is defined for installing the packages required for development and software testing, <code>benchmarking</code> is for installing the packages required for benchmarking against differnt TMM software programs, and <code>examples</code> is needed for running the examples included in the <code>examples</code> directory. One of these modes can be used by commanding <code>pip install .[dev]</code>, <code>pip install .[benchmarking]</code>, or <code>pip install .[examples]</code>.
 
 Examples
 ========
 
 A collection of examples in the <code>examples</code> directory exhibits various use cases and capabilities of our software. We provide the following examples:
 
-1. [Reflection Spectra over Wavelengths Varying Incidence Angles](https://github.com/JaxLayerLumos/JaxLayerLumos/blob/main/examples/angle-variation.ipynb)
-2. [Color Conversion](https://github.com/JaxLayerLumos/JaxLayerLumos/blob/main/examples/color-conversion.ipynb)
-3. [Color Exploration with Thin-Film Structures](https://github.com/JaxLayerLumos/JaxLayerLumos/blob/main/examples/color-exploration.ipynb)
-4. [Gradient Computation](https://github.com/JaxLayerLumos/JaxLayerLumos/blob/main/examples/gradient-computation.ipynb)
-5. [Visualization of Light Sources](https://github.com/JaxLayerLumos/JaxLayerLumos/blob/main/examples/light-source-visualization.ipynb)
+1. `Reflection Spectra over Wavelengths Varying Incidence Angles <https://github.com/JaxLayerLumos/JaxLayerLumos/blob/main/examples/angle-variation.ipynb>`_
+2. `Color Conversion <https://github.com/JaxLayerLumos/JaxLayerLumos/blob/main/examples/color-conversion.ipynb>`_
+3. `Color Exploration with Thin-Film Structures <https://github.com/JaxLayerLumos/JaxLayerLumos/blob/main/examples/color-exploration.ipynb>`_
+4. `Gradient Computation <https://github.com/JaxLayerLumos/JaxLayerLumos/blob/main/examples/gradient-computation.ipynb>`_
[…]
