+++ docs/about/about_jaxlayerlumos.rst
@@ -82,7 +82,7 @@ Citation
 .. code-block:: latex
 
     @misc{LiM2024jaxlayerlumos,
-        title={{JaxLayerLumos}: A {JAX}-based Efficient Transfer-Matrix Method Framework for Optical Simulations},
+        title={{JaxLayerLumos}: A {JAX}-based Differentiable Optical and Radio Frequency Simulator for Multilayer Structures},
         author={Li, Mingxuan and Kim, Jungtaek and Leu, Paul W.},
         howpublished={\url{https://doi.org/10.5281/zenodo.12602789}},
         year={2024}
