+++ docs/about/about_jaxlayerlumos.rst
@@ -76,6 +76,13 @@ License
 
 JaxLayerLumos is released under the `MIT License <https://github.com/JaxLayerLumos/JaxLayerLumos/blob/main/LICENSE>`_, promoting open and unrestricted access to software for academic and commercial use.
 
+Acknowledgments
+===============
+
+- We sincerely thank all contributors and users for your support and feedback.
+- This work is supported by the `Center for Materials Data Science for Reliability and Degradation (MDS-Rely) <https://mds-rely.org>`_, which is the `Industry-University Cooperative Research Center (IUCRC) <https://iucrc.nsf.gov>`_ of `National Science Foundation <https://www.nsf.gov>`_.
+- The University of Pittsburgh, Case Western Reserve University, and Carnegie Mellon University are participating institutions in MDS-Rely.
+
 Citation
 ========
 
