### 2025-09-07 JaxLayerLumos/JaxLayerLumos#25 @jungtaekkim
Thank you for your valuable comment.

We have added a footnote about this issue.
