### 2025-09-07 JaxLayerLumos/JaxLayerLumos#26 @brandondube
The docs homepage https://jaxlayerlumos.readthedocs.io/en/joss/index.html

Contains a link to https://jaxlayerlumos.readthedocs.io/en/joss/about/about_jaxlayerlumos.html

which, imo, is the "real" docs site; having the left table of contents and much more content, etc.  Perhaps it would make more sense to re-arrange the docs home page at this time
