+++ docs/about/software_testing.rst
@@ -1,0 +1,4 @@
+Software Testing and Test Automation
+####################################
+
+We provide a variety of test files in the ``tests`` directory. Before running the test files, the required packages should be installed by using ``pip install .[dev]``. They can be run by commanding ``pytest tests/``. Moreover, these test files are automatically tested via GitHub Actions, of which the configuration is defined in ``.github/workflows/pytest.yml``.
