+++ docs/about/examples.rst
@@ -1,0 +1,19 @@
+Examples
+########
+
+A collection of examples in the ``examples`` directory exhibits various use cases and capabilities of our software. We provide the following examples:
+
+1. `Reflection Spectra over Wavelengths Varying Incidence Angles <https://github.com/JaxLayerLumos/JaxLayerLumos/blob/main/examples/angle-variation.ipynb>`_
+2. `Color Conversion <https://github.com/JaxLayerLumos/JaxLayerLumos/blob/main/examples/color-conversion.ipynb>`_
+3. `Color Exploration with Thin-Film Structures <https://github.com/JaxLayerLumos/JaxLayerLumos/blob/main/examples/color-exploration.ipynb>`_
+4. `Gradient Computation <https://github.com/JaxLayerLumos/JaxLayerLumos/blob/main/examples/gradient-computation.ipynb>`_
+5. `Visualization of Light Sources <https://github.com/JaxLayerLumos/JaxLayerLumos/blob/main/examples/light-source-visualization.ipynb>`_
+
+6. `Plotting of Optical Constants <https://github.com/JaxLayerLumos/JaxLayerLumos/blob/main/examples/n-k-extrapolation.ipynb>`_
+7. `Thin-Film Structure Optimization with Bayesian Optimization <https://github.com/JaxLayerLumos/JaxLayerLumos/blob/main/examples/optimization-bayeso.ipynb>`_
+8. `Thin-Film Structure Optimization with DoG Optimizer <https://github.com/JaxLayerLumos/JaxLayerLumos/blob/main/examples/optimization-dog.ipynb>`_
+9. `Reflection Spectra over Frequencies for Radar Design <https://github.com/JaxLayerLumos/JaxLayerLumos/blob/main/examples/radar-design.ipynb>`_
+10. `Analysis of Solar Cells <https://github.com/JaxLayerLumos/JaxLayerLumos/blob/main/examples/solar-cell-analysis.ipynb>`_
+
+11. `Transmission Spectra over Wavelengths Varying Thicknesses <https://github.com/JaxLayerLumos/JaxLayerLumos/blob/main/examples/thickness-variation.ipynb>`_
+12. `Triple Junction Solar Cells <https://github.com/JaxLayerLumos/JaxLayerLumos/blob/main/examples/triple-junction-solar-cells.ipynb>`_
