+++ docs/about/license.rst
@@ -1,0 +1,4 @@
+License
+#######
+
+JaxLayerLumos is released under the `MIT License <https://github.com/JaxLayerLumos/JaxLayerLumos/blob/main/LICENSE>`_, promoting open and unrestricted access to software for academic and commercial use.
