+++ docs/index.rst
@@ -16,10 +16,6 @@ Documentation for JaxLayerLumos
    about/installation
    about/examples
    about/software_testing
-   about/supported_materials
-   about/contributing_guidelines
-   about/license
-   about/acknowledgments
 
 .. toctree::
     :maxdepth: 2
