+++ docs/python_api/jaxlayerlumos.rst
@@ -1,6 +1,12 @@
 jaxlayerlumos
 =============
 
+.. automodule:: jaxlayerlumos
+    :members:
+    :private-members:
+    :undoc-members:
+    :show-inheritance:
+
 .. toctree::
     jaxlayerlumos.jaxlayerlumos
     jaxlayerlumos.utils_layers
