+++ docs/index.rst
@@ -1,15 +1,23 @@
-Documentation for JaxLayerLumos
-===============================
+JaxLayerLumos: A JAX-based Differentiable Optical and Radio Frequency Simulator for Multilayer Structures
+=========================================================================================================
 
-.. automodule:: jaxlayerlumos
-    :members:
-    :private-members:
-    :undoc-members:
-    :show-inheritance:
+.. image:: _static/img/jaxlayerlumos.jpg
+    :width: 400
+    :align: center
+    :alt: jaxlayerlumos
+
+JaxLayerLumos is open-source transfer-matrix method (TMM) software designed for scientists, engineers, and researchers in optics and photonics. It provides a powerful yet intuitive interface for calculating the reflection and transmission (RT) of light through multi-layer optical structures. By inputting the refractive index, thickness of each layer, and the frequency vector, users can analyze how light interacts with layered materials, including the option to adjust for incidence angles.
+
+.. image:: _static/img/TMM.jpg
+    :width: 400
+    :align: center
+    :alt: TMM
+
+Our mission is to offer a lightweight, flexible, and fast alternative to commercial software, enabling users to perform complex optical simulations with ease. JaxLayerLumos is built with performance and usability in mind, facilitating the exploration of optical phenomena in research and development settings.
 
 .. toctree::
     :maxdepth: 1
-    :caption: JaxLayerLumos:
+    :caption: About JaxLayerLumos:
 
     about/overview
     about/features
