+++ docs/about/supported_materials.rst
@@ -1,12 +1,12 @@
 Supported Materials
 ###################
 
-.. image:: _static/img/Ag.jpg
+.. image:: _static/img/Ag.png
     :width: 400
     :align: center
     :alt: Ag
 
-.. image:: _static/img/TiO2.jpg
+.. image:: _static/img/TiO2.png
     :width: 400
     :align: center
     :alt: TiO2
