+++ docs/about/examples.rst
@@ -1,7 +1,7 @@
 Examples
 ########
 
-.. image:: _static/img/applications.png
+.. image:: ../_static/img/applications.png
     :width: 400
     :align: center
     :alt: applications
