+++ docs/about/features.rst
@@ -1,8 +1,0 @@
-Features
-########
-
-- Gradient Calculation: Calculates the gradients over any variables involved in RT, powered by JAX.
-- Flexibility: Accommodates a wide range of materials and structures by allowing users to specify complex refractive indices, layer thicknesses, and frequency vectors.
-- Angle of Incidence Support: Expands simulation capabilities to include angled light incidence, providing more detailed analysis for advanced optical designs.
-- Open Source and Community-Driven: Encourages contributions and feedback from the community, ensuring continuous improvement and innovation.
-- Comprehensive Material Database: Includes a growing database of materials with their optical properties, streamlining the simulation setup process.
