+++ docs/index.rst
@@ -15,11 +15,19 @@ JaxLayerLumos is open-source transfer-matrix method (TMM) software designed for 
 
 Our mission is to offer a lightweight, flexible, and fast alternative to commercial software, enabling users to perform complex optical simulations with ease. JaxLayerLumos is built with performance and usability in mind, facilitating the exploration of optical phenomena in research and development settings.
 
+Features
+--------
+
+- Gradient Calculation: Calculates the gradients over any variables involved in RT, powered by JAX.
+- Flexibility: Accommodates a wide range of materials and structures by allowing users to specify complex refractive indices, layer thicknesses, and frequency vectors.
+- Angle of Incidence Support: Expands simulation capabilities to include angled light incidence, providing more detailed analysis for advanced optical designs.
+- Open Source and Community-Driven: Encourages contributions and feedback from the community, ensuring continuous improvement and innovation.
+- Comprehensive Material Database: Includes a growing database of materials with their optical properties, streamlining the simulation setup process.
+
 .. toctree::
     :maxdepth: 1
     :caption: About JaxLayerLumos:
 
-    about/features
     about/installation
     about/examples
     about/software_testing
