+++ docs/about/acknowledgments_citation.rst
@@ -1,6 +1,21 @@
-Acknowledgments
-###############
+Acknowledgments and Citation
+############################
+
+Acknowledgments:
+----------------
 
 - We sincerely thank all contributors and users for your support and feedback.
 - This work is supported by the `Center for Materials Data Science for Reliability and Degradation (MDS-Rely) <https://mds-rely.org>`_, which is the `Industry-University Cooperative Research Center (IUCRC) <https://iucrc.nsf.gov>`_ of `National Science Foundation <https://www.nsf.gov>`_.
 - The University of Pittsburgh, Case Western Reserve University, and Carnegie Mellon University are participating institutions in MDS-Rely.
+
+Citation:
+---------
+
+.. code-block:: latex
+
+    @misc{LiM2024jaxlayerlumos,
+        title={{JaxLayerLumos}: A {JAX}-based Differentiable Optical and Radio Frequency Simulator for Multilayer Structures},
+        author={Li, Mingxuan and Kim, Jungtaek and Leu, Paul W.},
+        howpublished={\url{https://doi.org/10.5281/zenodo.12602789}},
+        year={2024}
+    }
