+++ docs/about/citation.rst
@@ -1,11 +1,0 @@
-Citation
-########
-
-.. code-block:: latex
-
-    @misc{LiM2024jaxlayerlumos,
-        title={{JaxLayerLumos}: A {JAX}-based Differentiable Optical and Radio Frequency Simulator for Multilayer Structures},
-        author={Li, Mingxuan and Kim, Jungtaek and Leu, Paul W.},
-        howpublished={\url{https://doi.org/10.5281/zenodo.12602789}},
-        year={2024}
-    }
