### 2025-09-16 c9c29fd
Update paper.md
