### 2025-10-02 #8572 @zhubonan
Hi @Mil152 just one minor comment: we suggest using "United State of America" instead of the "USA" in the author affiliation part of the paper. Otherwise, it looks all fine to me. The next thing to do is to make a release of the software and archive it on Zenodo to get an DOI. Please make sure the release have the same title and authors as the paper.

Once it is done, please let me know the DOI and the version of the code. Thanks！
