### 2025-10-02 1d5129e
Update paper.md
