### 2025-10-02 #8572 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
