### 2025-10-08 #8572 @rkurchin
👋 Hi, congrats on being nearly done! As AEiC for JOSS I will now help to process this submission for acceptance in JOSS. Below are some final checks, some of which may require your attention:
- [ ] your contributing guidelines on the docs site say to read `CONTRIBUTING.md`, but this file doesn't appear to exist
- [ ] update Scikit-learn citation – see [here](https://scikit-learn.org/stable/about.html#citing-scikit-learn)
- [ ] "Mention of use" feels awkward as a section title. I might change it to "Potential applications" – you should also rephrase the start of the section to make clear that the citation don't literally use your software (obviously JAX didn't exist in 1993...), but rather are representative types of analyses that it would be capable of (I presume)
