### 2025-10-08 ee92adc
Update paper.bib
