+++ paper/paper.bib
@@ -1,20 +1,19 @@
 @software{jax2018github,
-  author = {James Bradbury and Roy Frostig and Peter Hawkins and Matthew James Johnson and Chris Leary and Dougal Maclaurin and George Necula and Adam Paszke and Jake Vander{P}las and Skye Wanderman-{M}ilne and Qiao Zhang},
-  title = {{JAX}: composable transformations of {P}ython+{N}um{P}y programs},
-  url = {http://github.com/jax-ml/jax},
-  version = {0.3.13},
-  year = {2018}
+  author={James Bradbury and Roy Frostig and Peter Hawkins and Matthew James Johnson and Chris Leary and Dougal Maclaurin and George Necula and Adam Paszke and Jake Vander{P}las and Skye Wanderman-{M}ilne and Qiao Zhang},
+  title={{JAX}: composable transformations of {P}ython+{N}um{P}y programs},
+  url={https://github.com/jax-ml/jax},
+  version={0.3.13},
+  year={2018}
 }
 
 @article{Wang:15apr,
-  author = {Wang, Baomin and Leu, Paul W.},
-  doi = {10.1016/j.nanoen.2014.10.040},
-  journal = {Nano Energy},
-  month = Apr,
-  pages = {226--232},
-  title = {High Index of Refraction Nanosphere Coatings for Light Trapping in Crystalline Silicon Thin Film Solar Cells},
-  volume = {13},
-  year = {2015}
+  author={Wang, Baomin and Leu, Paul W.},
+  doi={10.1016/j.nanoen.2014.10.040},
+  journal={Nano Energy},
+  pages={226--232},
+  title={High Index of Refraction Nanosphere Coatings for Light Trapping in Crystalline Silicon Thin Film Solar Cells},
+  volume={13},
+  year={2015}
 }
 
 @article{Gao:14aug,
@@ -40,14 +39,14 @@
 }
 
 @article{Li:23,
-  author = {Li, Mingxuan and Sinha, Sneh and Hannani, Sima and Walker, S. Brett and LeMieux, Melbs and Leu, Paul W.},
-  doi = {10.1021/acsaelm.2c01183},
-  journal = {ACS Applied Electronic Materials},
-  title = {Ink-Coated Silver Films on {PET} for Flexible, High Performance Electromagnetic Interference Shielding and Joule Heating},
+  author={Li, Mingxuan and Sinha, Sneh and Hannani, Sima and Walker, S. Brett and LeMieux, Melbs and Leu, Paul W.},
+  doi={10.1021/acsaelm.2c01183},
+  journal={ACS Applied Electronic Materials},
+  title={Ink-Coated Silver Films on {PET} for Flexible, High Performance Electromagnetic Interference Shielding and Joule Heating},
   volume={5},
   number={1},
   pages={173--180},
-  year = {2023}
+  year={2023}
 }
 
 @article{Li:22,
@@ -82,12 +81,11 @@
 }
 
 @software{tmmSbyrnes,
-  author = {Byrnes, Steven},
-  title = {{tmm}: Multilayer optical calculations},
-  year = {2012},
-  version = {0.2.0},
-  url = {https://github.com/sbyrnes321/tmm},
-  doi = {10.48550/arXiv.1603.02720}
+  author={Byrnes, Steven},
+  title={{tmm}: Multilayer optical calculations},
+  year={2012},
+  version={0.2.0},
+  url={https://github.com/sbyrnes321/tmm}
 }
 
 @article{luce2022tmm,
@@ -161,14 +159,14 @@
 }
 
 @article{pedregosa2011scikit,
[…]
