### 2025-10-11 f64ddf6
Update paper.md
