### 2025-10-13 8f957b4
Update paper.md
